# attached_assets/phmto-master/static/openapi.yaml
openapi: 3.0.3
info:
  title: Phmto API (Solana Wallet)
  version: "3.5.0"
  description: >
    API Flask pour gérer des projets & wallets Solana (devnet/mainnet).
    Inclut airdrop (devnet), transferts, mix, consolidation, import/export,
    et gestion de token avec création via Pump.fun.
servers:
  - url: http://localhost:8000
    description: Local
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/projects/{project_id}/wallets:
    parameters:
      - in: path
        name: project_id
        required: true
        schema: { type: string }
    get:
      tags: [Wallets]
      summary: Lister les wallets d’un projet
      parameters:
        - in: query
          name: with_balance
          schema: { type: string, enum: ["true","false"] }
      responses:
        "200":
          description: Liste de wallets
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  project_id: { type: string }
                  name: { type: string }
                  wallets:
                    type: array
                    items:
                      $ref: "#/components/schemas/WalletPublic"
    post:
      tags: [Wallets]
      summary: Générer N wallets dans un projet
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                n:
                  type: integer
                  minimum: 1
                  maximum: 1000
                  default: 1
      responses:
        "201":
          description: Wallets créés
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  created: { type: integer }
                  wallets:
                    type: array
                    items:
                      $ref: "#/components/schemas/WalletFull"
...
  /api/v1/projects/{project_id}/wallets/{wallet_id}:
    parameters:
      - in: path
        name: project_id
        required: true
        schema: { type: string }
      - in: path
        name: wallet_id
        required: true
        schema: { type: string }
    delete:
      tags: [Wallets]
      summary: Supprimer un wallet du projet (backup JSON)
      responses:
        "200": { description: Wallet supprimé }
        "404": { description: Introuvable }

  # =============== Wallets (globaux) ===============
  /api/v1/wallets/{wallet_id}:
    parameters:
      - in: path
        name: wallet_id
        required: true
        schema: { type: string }
    get:
      tags: [Wallets]
      summary: Détail complet d’un wallet (id, name, pub/priv, date, solde optionnel)
      parameters:
        - in: query
          name: with_balance
          schema: { type: string, enum: ["true","false"] }
      responses:
        "200":
          description: Détail du wallet
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  wallet:
                    $ref: "#/components/schemas/WalletFull"
        "404": { description: Introuvable }
    patch:
      tags: [Wallets]
      summary: Renommer un wallet par ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        "200": { description: OK }
        "404": { description: Introuvable }
...
  /api/v1/wallets/{address}/balance:
    parameters:
      - in: path
        name: address
        required: true
        schema: { type: string }
      - in: query
        name: cluster
        schema: { type: string, enum: [devnet, testnet, mainnet] }
      - in: query
        name: rpc
        schema: { type: string }
    get:
      tags: [Wallets]
      summary: Solde SOL d’une adresse
      responses:
        "200":
          description: Solde
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBalanceResponse"
        "400": { description: Requête invalide }

  # ===================== Transfers / Airdrop =====================
...
  /api/v1/wallets/{wallet_id}/transfer:
    parameters:
      - in: path
        name: wallet_id
        required: true
        schema: { type: string }
      - in: query
        name: cluster
        schema: { type: string, enum: [devnet, testnet, mainnet] }
      - in: query
        name: rpc
        schema: { type: string }
    post:
      tags: [Transfers]
      summary: Transférer des SOL (wallet_id → pubkey destination)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [recipient_pubkey, amount_sol]
              properties:
                recipient_pubkey: { type: string }
                amount_sol:
                  type: number
                  format: float
      responses:
        "200":
          description: Transfert effectué
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResult"
        "400": { description: Requête invalide }
        "404": { description: Introuvable }
        "500": { description: Erreur interne }
  /api/v1/wallets/mix:
    post:
      tags: [Transfers]
      summary: Mixer des SOL entre wallets (random/roundrobin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [wallet_ids]
              properties:
                wallet_ids:
                  type: array
                  items: { type: string }
                strategy:
                  type: string
                  enum: [random, roundrobin]
                  default: random
      responses:
        "200":
          description: Historique des transferts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MixResponse"
        "400": { description: Requête invalide }
        "404": { description: Introuvable }
        "500": { description: Erreur mix }
  /api/v1/wallets/consolidate/{target_wallet_id}:
    parameters:
      - in: path
        name: target_wallet_id
        required: true
        schema: { type: string }
    post:
      tags: [Transfers]
      summary: Consolider les soldes vers un wallet cible
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                project_id: { type: string }
                min_reserve_sol:
                  type: number
                  format: float
                  default: 0.00001
      responses:
        "200":
          description: Historique + ignorés
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsolidateResponse"
        "400": { description: Requête invalide }
        "404": { description: Introuvable }

  # ===================== Tokens =====================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    WalletPublic:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        address: { type: string }
        created_at: { type: string }
        balance_sol: { type: number, format: float }
    WalletFull:
      allOf:
        - $ref: "#/components/schemas/WalletPublic"
        - type: object
          properties:
            private_key: { type: string, description: "clé privée base58 (en clair ici)" }
            private_key_json_64:
              type: array
              items: { type: integer }
...
    WalletBalanceResponse:
      type: object
      properties:
        ok: { type: boolean }
        address: { type: string }
        balance_sol: { type: number, format: float }
        rpc_url: { type: string }
        cluster: { type: string }
    TransferResult:
      type: object
      properties:
        ok: { type: boolean }
        from_wallet_id: { type: string }
        from_address: { type: string }
        to_address: { type: string }
        amount_sol: { type: number, format: float }
        signature: { type: string }
        rpc_url: { type: string }
    MixHistoryItem:
      type: object
      properties:
        from_wallet_id: { type: string }
        from_address: { type: string }
        to_address: { type: string }
        amount_sol: { type: number, format: float }
        signature: { type: string }
    MixResponse:
      type: object
      properties:
        ok: { type: boolean }
        strategy: { type: string }
        transfers:
          type: array
          items:
            $ref: "#/components/schemas/MixHistoryItem"
    ConsolidateHistoryItem:
      type: object
      properties:
        from_wallet_id: { type: string }
        from_address: { type: string }
        to_wallet_id: { type: string }
        to_address: { type: string }
        amount_sol: { type: number, format: float }
        signature: { type: string }
    ConsolidateResponse:
      type: object
      properties:
        ok: { type: boolean }
        target: { type: string }
        transfers:
          type: array
          items:
            $ref: "#/components/schemas/ConsolidateHistoryItem"
        skipped:
          type: array
          items:
            type: object
            properties:
              wallet_id: { type: string }
              reason: { type: string }
...